% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pen.R
\name{addConstraint}
\alias{addConstraint}
\title{Add constraint to optiProject}
\usage{
addConstraint(optiProject, state, limit, time, pen_values, gamma = 100)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{state}{State in the structural model to be penalized.}

\item{limit}{The limit, under/above which the state gets penalized.}

\item{time}{Time-point of penalization (one numeric value).}

\item{pen_values}{What values should be penalized; "below" if limit is the lower and "above" if
limit is the upper limit.}

\item{gamma}{Penalization strength.}
}
\description{
Add a constraint, i.e., penalizing state-values below or above a certain limit, 
at a specific time-point.
}
\details{
Constraint at a specific time-points may be useful if only a specific time-point is of
interest, e.g., trough-concentrations. For broader application, see \link[=func]{addConstraintAUC}.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2") \%>\%
                  addConstraint("R",limit=30,pen_time = 50,pen_values="below")
}
}
\author{
Dominic Bräm
}
