% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pen_AUC.R
\name{addConstraintAUC}
\alias{addConstraintAUC}
\title{Add constraint AUC to optiProject}
\usage{
addConstraintAUC(
  optiProject,
  state,
  limit,
  pen_time,
  eval_time,
  pen_values,
  gamma = 100
)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{state}{State in the structural model to be penalized.}

\item{limit}{The limit, under/above which the state gets penalized.}

\item{pen_time}{Time-range when penalization should be applied (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{eval_time}{Time-point at which the constraint AUC should be evaluated.}

\item{pen_values}{What values should be penalized; "below" if limit is the lower and "above" if
limit is the upper limit.}

\item{gamma}{Penalization strength.}
}
\description{
Add a constraint, i.e., penalizing state-values below or above a certain limit, 
over a time-frame.
}
\details{
The area between limit and state is calculated if state-values are above or below 
the limit, within the defined penalization time-frame.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2") \%>\%
                  addConstraint("R",limit=30,pen_time = c(0,200),
                                eval_time=200,pen_values="below",gamma=1000)
}
}
\author{
Dominic Bräm
}
