% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder.R
\name{addDoseLevel}
\alias{addDoseLevel}
\title{Add dose level to optiProject-object}
\usage{
addDoseLevel(
  optiProject,
  time,
  n_dose = 1,
  ini_est = 1,
  dose_interval = NULL,
  adm = NULL
)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{time}{Time of first dose for this dose level.}

\item{n_dose}{Number of doses for this dose level.}

\item{ini_est}{Initial estimate for this dose level.}

\item{dose_interval}{If \code{n_dose>1}, a dosing interval is required.}

\item{adm}{If different administration IDs are present in the structural model,
the administration ID of this dose level must be defined.}
}
\description{
Add a dose level to be estimated to a optiProject-object
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxModel("../inst/oral1_1cpt_IndirectModelInhibitionKin_TlagkaVClR0koutImaxIC50.txt") \%>\%
                  addPopParms("../inst/warfarinPKPD_project/populationParameters.txt") \%>\%
                  addDoseLevel(time=0,ini_est = 100)
}
}
\author{
Dominic Bräm
}
