% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder.R
\name{addIndParms}
\alias{addIndParms}
\title{Add individual estimates to optiProject-object}
\usage{
addIndParms(optiProject, pmx_parms, phi_file = NULL)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{pmx_parms}{Either a file name (estimatedIndividualParameters.txt for "Monolix" or results file for
"NONMEM") containing population estimates, or a data frame with a column "id" for subject IDs and columns 
for each parameter (if "Monolix", the explicit individual parameter, if "NONMEM", THETAs and ETAs).}

\item{phi_file}{If software is "NONMEM" and pmx_parms not a data frame, a .phi file of the fitted model
must be given to extract individual ETA estimates.}
}
\description{
If not an already fitted pharmacometric model is added through addPmxRun, explicit individual
parameter estimates can be added to the optiProject-object with this function.
}
\details{
If software ist "NONMEM", the parameter names must be of type THETAX and ETAX with X the number 
in THETA(X) and ETA(X) in the control file. This function is not applicable for nlmixr2 as
a fitted nlmixr2 object can directly added with \link[=func]{addPmxRun}.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxModel("../inst/oral1_1cpt_IndirectModelInhibitionKin_TlagkaVClR0koutImaxIC50.txt") \%>\%
                  addIndParms("../inst/warfarinPKPD_project/IndividualParameters/estimatedIndividualParameters.txt")
}
}
\author{
Dominic Bräm
}
