% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder.R
\name{addPmxModel}
\alias{addPmxModel}
\title{Add structural model to optiProject-object}
\usage{
addPmxModel(optiProject, pmx_model)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{pmx_model}{(Path/)File name of structural model or rxode2 style model function 
to be added to the optiProject-object.If software is "Monolix", a \emph{.txt} file is 
required (as in a Monolix project), if software is "NONMEM", the NONMEM control file 
name can be added, if software is "nlmixr2", a rxode2 style model function that would be
used in the nlmixr2 fit can be given.}
}
\description{
If not an already fitted pharmacometric model is added through addPmxRun, an explicit structural
model can be added to the optiProject-object with this function.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxModel("../inst/oral1_1cpt_IndirectModelInhibitionKin_TlagkaVClR0koutImaxIC50.txt")
}
}
\author{
Dominic Bräm
}
