% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder.R
\name{addPmxRun}
\alias{addPmxRun}
\title{Adding an already fitted pharmacometric model}
\usage{
addPmxRun(
  optiProject,
  pmx_file,
  lixoft_path = NULL,
  res_file = NULL,
  phi_file = NULL
)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{pmx_file}{An already fitted pharmacometric model, i.e., a \emph{.mlxtran} file
if optiProject-object is initialized with "Monolix", or a NONMEM control file if initialized
with "NONMEM".}

\item{lixoft_path}{If initialized with "Monolix" and the structural model used in the \emph{.mlxtran} file is
a model from the library, \code{lixoft_path} is needed. Standard is "C:/ProgramData/Lixoft/MonolixSuiteXXXX".}

\item{res_file}{A NONMEM results file if optiProject is initialized with "NONMEM" (in order
to extract population estimates).}

\item{phi_file}{A NONMEM .phi file if optiProject is initialized with "NONMEM" (in order
to extract individual estimates).}
}
\description{
Adding an already fitted pharmacometric model to the optiProject-object, including
structural model, population and individual parameter estimates.
}
\examples{
\dontrun{

warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2")
}
}
\author{
Dominic Bräm
}
