% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder.R
\name{addPopParms}
\alias{addPopParms}
\title{Add population estimates to optiProject-object}
\usage{
addPopParms(optiProject, pmx_parms)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{pmx_parms}{Either a file name (populationEstimates.txt for "Monolix" or results file for
"NONMEM") containing population estimates, or a data frame with a column "parameter" for parameter
names and "value" for the population estimates.}
}
\description{
If not an already fitted pharmacometric model is added through addPmxRun, explicit population
parameter estimates can be added to the optiProject-object with this function.
}
\details{
If software ist "NONMEM" and a data frame is given, the parameter names must be of type THETAX and ETAX with X the number 
in THETA(X) and ETA(X) in the control file. This function is not applicable for nlmixr2 as
a fitted nlmixr2 object can directly added with \link[=func]{addPmxRun}.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxModel("../inst/oral1_1cpt_IndirectModelInhibitionKin_TlagkaVClR0koutImaxIC50.txt") \%>\%
                  addPopParms("../inst/warfarinPKPD_project/populationParameters.txt")
}
}
\author{
Dominic Bräm
}
