% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pen.R
\name{addSecondary}
\alias{addSecondary}
\title{Add secondary target to optiProject}
\usage{
addSecondary(optiProject, state, pen_values, time, gamma = 1)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{state}{State in the structural model to be penalized.}

\item{pen_values}{What values should be penalized; "high" if high state.values and "low" if
low state-values should be penalized.}

\item{time}{Time-point of penalization (one numeric value).}

\item{gamma}{Penalization strength.}
}
\description{
Add a secondary target, i.e., penalizing low or high state-values, 
at a specific time-point.
}
\details{
Usually, secondary targets are required when a constraint is applied. 
Secondary targets at a specific time-points may be useful if only a specific time-point is of
interest, e.g., trough-concentrations. For broader application, see \link[=func]{addSecondaryAUC}.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2") \%>\%
                  addConstraint("R",limit = 30,time = 50,pen_values="below") \%>\%
                  addSecondary("Cc",pen_values = "low",time = 30)
}
}
\author{
Dominic Bräm
}
