% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pen_AUC.R
\name{addSecondaryAUC}
\alias{addSecondaryAUC}
\title{Add secondary target AUC to optiProject}
\usage{
addSecondaryAUC(optiProject, state, pen_values, pen_time, eval_time, gamma = 1)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{state}{State in the structural model to be penalized.}

\item{pen_values}{What values should be penalized; "high" if high state.values and "low" if
low state-values should be penalized.}

\item{pen_time}{Time-range when penalization should be applied (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{eval_time}{Time-point at which the constraint AUC should be evaluated.}

\item{gamma}{Penalization strength.}
}
\description{
Add a secondary target, i.e., penalizing low or high state-AUC, 
over a time-frame.
}
\details{
The area under the state-curve is calculated and either high or low AUCs are penalized.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2") \%>\%
                  addConstraintAUC("R",limit = 30,pen_time = c(0,50), eval_time = 50, pen_values="below") \%>\%
                  addSecondaryAUC("Cc",pen_values = "low",pen_time = c(0,50), eval_time = 50)
}
}
\author{
Dominic Bräm
}
