% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pen.R
\name{addTargetFun}
\alias{addTargetFun}
\title{Add a target function to optiProject}
\usage{
addTargetFun(optiProject, state, target_fun, time, gamma = 1)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{state}{State in the structural model to be penalized.}

\item{target_fun}{Target function of the state.}

\item{time}{Time-point of penalization (one numeric value).}

\item{gamma}{Penalization strength.}
}
\description{
Add a target function, i.e., penalizing state-values different from the target function values, 
at a specific time-point.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2") \%>\%
                  addTargetFun("R",target_fun = "100 \* exp(-0.2 \* t)", time = 50)
}
}
\author{
Dominic Bräm
}
