% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pen_AUC.R
\name{addTargetFunAUC}
\alias{addTargetFunAUC}
\title{Add AUC of divergence from target function to optiProject}
\usage{
addTargetFunAUC(optiProject, state, target_fun, pen_time, eval_time, gamma = 1)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{state}{State in the structural model to be penalized.}

\item{target_fun}{Target function for the state.}

\item{pen_time}{Time-range when penalization should be applied (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{eval_time}{Time-point at which the constraint AUC should be evaluated.}

\item{gamma}{Penalization strength.}
}
\description{
Add a target function, i.e., penalizing AUC from state-values different from this function, 
over a time-frame.
}
\details{
The area between state-curve and target function is calculated and penalized.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2") \%>\%
                  addTargetFunAUC("R",target_fun = "100 \* exp(-0.2 \* t)", pen_time = c(0,30),
                                    eval_time = 50)
}
}
\author{
Dominic Bräm
}
