% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pen.R
\name{addTargetValue}
\alias{addTargetValue}
\title{Add a target value to optiProject}
\usage{
addTargetValue(optiProject, state, target_value, time, gamma = 1)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{state}{State in the structural model to be penalized.}

\item{target_value}{Target value of the state.}

\item{time}{Time-point of penalization (one numeric value).}

\item{gamma}{Penalization strength.}
}
\description{
Add a target value, i.e., penalizing state-values different from this value, 
at a specific time-point.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2") \%>\%
                  addTargetValue("R",target_value = 20, time = 50)
}
}
\author{
Dominic Bräm
}
