% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pen_AUC.R
\name{addTargetValueAUC}
\alias{addTargetValueAUC}
\title{Add AUC of divergence from target value to optiProject}
\usage{
addTargetValueAUC(
  optiProject,
  state,
  target_value,
  pen_time,
  eval_time,
  gamma = 100
)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{state}{State in the structural model to be penalized.}

\item{target_value}{Target value of the state.}

\item{pen_time}{Time-range when penalization should be applied (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{eval_time}{Time-point at which the constraint AUC should be evaluated.}

\item{gamma}{Penalization strength.}
}
\description{
Add a target value, i.e., penalizing AUC from state-values different from this value, 
over a time-frame.
}
\details{
The area between state-curve and target value is calculated and penalized.
}
\examples{
\dontrun{
warf_proj <- newOptiProject("Monolix") \%>\%
                  addPmxRun("../inst/warfarinPKPD_project.mlxtran",
                            lixoft_path = "C:/ProgramData/Lixoft/MonolixSuite2021R2") \%>\%
                  addTargetValueAUC("R",target_value = 20, pen_time = c(30,50),
                                    eval_time = 50)
}
}
\author{
Dominic Bräm
}
