% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{ConstraintDataGen}
\alias{ConstraintDataGen}
\title{Generate penalization data}
\usage{
ConstraintDataGen(constraints)
}
\arguments{
\item{constraints}{Penalization constraints/targets in an optiProject-object.}
}
\description{
Generate penalization data from the constraints information of an optiProject-object
}
\author{
Dominic Bräm
}
