% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{DataGen}
\alias{DataGen}
\title{Generate data for an optiProject}
\usage{
DataGen(
  optiProject,
  pop = FALSE,
  sim = FALSE,
  n_id = 1,
  parametrisation = "log",
  lower_parm_limits = NULL,
  upper_parm_limits = NULL
)
}
\arguments{
\item{optiProject}{A optiProject-object with a model, penalizations, dose levels, and parameters.
For requirements regarding penalizations, dose levels, and parameters.}

\item{pop}{Whether multiple subjects should be used (TRUE), either from added individual parameters or
simulated from population parameters, or if only one subject with population parameters should be used
(FALSE).}

\item{sim}{If \code{pop=TRUE}, whether individuals should be simulated (TRUE) or added individual
parameters should be used (FALSE).}

\item{n_id}{If \code{pop=TRUE} and \code{sim=TRUE}, number of individuals to be simulated.}

\item{parametrisation}{If a Monolix-optiProject was created and a population of individuals is simulated
(\code{pop=TRUE} and \code{sim=TRUE}), the parametrisation of the parameters in the Monolix model
must be given. Must be in "log","normal",or "logit". If all parameters have same parametrisation,
a single string can be given. If different parametrisations for different parameters are required, a vector
with parametrisation for each parameter is required. Default is "log" for all parameters.}

\item{lower_parm_limits}{If one or multiple parameters in the model are "logit" distributed and individuals
are simulated, the lower limit(s) for the "logit"-distributed parameter(s).}

\item{upper_parm_limits}{If one or multiple parameters in the model are "logit" distributed and individuals
are simulated, the upper limit(s) for the "logit"-distributed parameter(s).}

\item{iiv}{Whether doses should include inter-individual variability (TRUE) or not (FALSE).}
}
\description{
Transforming targets, dose levels, and parameters of an optiProject to a data file
}
\author{
Dominic Bräm
}
