% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{DoseDataGen}
\alias{DoseDataGen}
\title{Generate dose data}
\usage{
DoseDataGen(doses)
}
\arguments{
\item{doses}{Dose levels in an optiProject-object added through \code{addDoseLevel}.}
}
\description{
Generate dose data from the dose level information of an optiProject-object
}
\author{
Dominic Bräm
}
