% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{eta_extractor_nlmixr}
\alias{eta_extractor_nlmixr}
\title{Etas extractor for nlmixr2}
\usage{
eta_extractor_nlmixr(fit_obj)
}
\arguments{
\item{fit_obj}{Fitted object with nlmixr2.}
}
\description{
Extract estimated individual Etas from nlmixr2 fit.
}
\author{
Dominic Bräm
}
