% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{eta_extractor_nm}
\alias{eta_extractor_nm}
\title{Individual estimates extractor for NONMEM}
\usage{
eta_extractor_nm(phi_file)
}
\arguments{
\item{phi_file}{NONMEM .phi file}
}
\description{
Extracts individual Etas from a NONMEM .phi file.
}
\author{
Dominic Bräm
}
