% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{iparm_extractor_mlx}
\alias{iparm_extractor_mlx}
\title{Individual estimates extractor for Monolix}
\usage{
iparm_extractor_mlx(iparm_file)
}
\arguments{
\item{iparm_file}{File of estimated individual parameters.}
}
\description{
Extracts individual estimates (EBEs) from a Monolix run.
}
\author{
Dominic Bräm
}
