% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_helpers.R
\name{mlx_admin_checker}
\alias{mlx_admin_checker}
\title{Check for administration macros}
\usage{
mlx_admin_checker(model)
}
\arguments{
\item{model}{Monolix structural model.}
}
\description{
The optiDose package only works if administration macros, i.e., depot, absorption,
iv, or pkmodel were used in the structural model. This function checks whether this is 
the case and if multiple administration IDs are present in the dose levels, whether 
these were already defined in the structural model.
}
\author{
Dominic Bräm
}
