% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_helpers.R
\name{mlx_header_types}
\alias{mlx_header_types}
\title{Identify header types}
\usage{
mlx_header_types(data)
}
\arguments{
\item{data}{Dataframe of the optiProject, generated through DataGen.}
}
\description{
Defines header types of optiProject-data.
}
\author{
Dominic Bräm
}
