% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{mlxConstraint}
\alias{mlxConstraint}
\title{Generate Monolix code for constraint}
\usage{
mlxConstraint(state, limit, time, pen_values, n_constraint, gamma)
}
\arguments{
\item{state}{State in the structural model to be penalized.}

\item{limit}{The limit, under/above which the state gets penalized.}

\item{time}{Time-point of penalization (one numeric value).}

\item{pen_values}{What values should be penalized; "below" if limit is the lower and "above" if
limit is the upper limit.}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}
}
\description{
Generate Monolix code for constraint, called by \link[=func]{addConstraint} if
software is "Monolix".
}
\author{
Dominic Bräm
}
