% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaries.R
\name{mlxSecondary}
\alias{mlxSecondary}
\title{Generate Monolix code for secondary target}
\usage{
mlxSecondary(state, time, pen_values, n_constraint, gamma, decay_high = 0.01)
}
\arguments{
\item{state}{State in the structural model to be penalized.}

\item{time}{Time-point of penalization (one numeric value).}

\item{pen_values}{What values should be penalized; "high" if high state.values and "low" if
low state-values should be penalized.}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}

\item{decay_high}{Rate of penalization-decay with increasing state-values.}
}
\description{
Generate Monolix code for secondary target, called by \link[=func]{addSecondary} if
software is "Monolix".
}
\details{
Penalization function is just \emph{state} if high state-values are penalized, and
exp(-0.01 * state) if low state-values are penalized.
}
\author{
Dominic Bräm
}
