% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_setup.R
\name{mlxSetAll}
\alias{mlxSetAll}
\title{Set up the Monolix optiProject}
\usage{
mlxSetAll(
  optiProject,
  name,
  save_path,
  pop,
  iiv,
  sim,
  n_id,
  parametrisation,
  lower_parm_limits,
  upper_parm_limits,
  const_err,
  create_mlx
)
}
\arguments{
\item{optiProject}{A optiProject-object with a model, penalizations, dose levels, and parameters.
For requirements regarding penalizations, dose levels, and parameters.}

\item{name}{Name under which optiProject will be saved. .mlxtran will be added to name.}

\item{save_path}{Path, where optiProject should be saved.}

\item{pop}{Whether multiple subjects should be used (TRUE), either from added individual parameters or
simulated from population parameters, or if only one subject with population parameters should be used
(FALSE).}

\item{iiv}{iiv Whether doses should include inter-individual variability (TRUE) or not (FALSE).}

\item{sim}{If \code{pop=TRUE}, whether individuals should be simulated (TRUE) or added individual
parameters should be used (FALSE).}

\item{n_id}{If \code{pop=TRUE} and \code{sim=TRUE}, number of individuals to be simulated.}

\item{parametrisation}{If a Monolix-optiProject was created and a population of individuals is simulated
(\code{pop=TRUE} and \code{sim=TRUE}), the parametrisation of the parameters in the Monolix model
must be given. Must be in "log","normal",or "logit". If all parameters have same parametrisation,
a single string can be given. If different parametrisations for different parameters are required, a vector
with parametrisation for each parameter is required. Default is "log" for all parameters.}

\item{lower_parm_limits}{If one or multiple parameters in the model are "logit" distributed and individuals
are simulated, the lower limit(s) for the "logit"-distributed parameter(s).}

\item{upper_parm_limits}{If one or multiple parameters in the model are "logit" distributed and individuals
are simulated, the upper limit(s) for the "logit"-distributed parameter(s).}

\item{const_err}{If doses are estimated with inter-individual variability for a population of individual
subjects, \code{const_err} defines the balance between fulfilling the objectives (constraints, target values etc.)
and having the doses log-normal distributed.}

\item{create_mlx}{If optiProject-software is "Monolix", whether directly a Monolix project (.mlxtran file) should
be created.}
}
\description{
Function to generate data and model from an optiProject-object
}
\author{
Dominic Bräm
}
