% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_setup.R
\name{mlxSetMlxtran}
\alias{mlxSetMlxtran}
\title{Set up the .mlxtran file}
\usage{
mlxSetMlxtran(
  optiProject,
  mlxtran_name,
  model_name,
  data_name,
  header_types,
  iiv = FALSE,
  const_err = 1e-04
)
}
\arguments{
\item{optiProject}{A optiProject-object with a model, penalizations, dose levels, and parameters.
For requirements regarding penalizations, dose levels, and parameters.}

\item{mlxtran_name}{Name under which the Monolix file should be saved. Must include .mlxtran}

\item{model_name}{Name of Monolix structural model file to be used.}

\item{data_name}{Name of data file to be used.}

\item{header_types}{Header types of columns in the data file.}

\item{iiv}{Whether doses should include inter-individual variability (TRUE) or not (FALSE).}

\item{const_err}{If doses are estimated with inter-individual variability for a population of individual
subjects, \code{const_err} defines the balance between fulfilling the objectives (constraints, target values etc.)
and having the doses log-normal distributed.}
}
\description{
Set up the .mlxtran file from an optiProject-object
}
\author{
Dominic Bräm
}
