% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_update_model.R
\name{mlxUpdateBioav}
\alias{mlxUpdateBioav}
\title{Update dosing macros}
\usage{
mlxUpdateBioav(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update dosing macros, such as depot, absorption, iv, and pkmodel, by adding 
bioavailability for dose levels.
}
\details{
Should be called after \code{mlxUpdatePK}
}
\author{
Dominic Bräm
}
