% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_update_model.R
\name{mlxUpdateEquation}
\alias{mlxUpdateEquation}
\title{Update Equation section}
\usage{
mlxUpdateEquation(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update EQUATION: section of the Monolix model by adding the penalization code of
the optiProject-object constraints
}
\details{
Should be called after \code{mlxUpdateBioav}
}
\author{
Dominic Bräm
}
