% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_update_model.R
\name{mlxUpdateInputs}
\alias{mlxUpdateInputs}
\title{Update input}
\usage{
mlxUpdateInputs(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update input to the Monolix model with F1-FX for dose levels, DOSE_NR as regressor to define 
when which dose level should be applied, and parmX_cov as regressor for parmX in the 
structural model.
}
\details{
Should be called after \code{mlxUpdateParmNames}
}
\author{
Dominic Bräm
}
