% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_update_model.R
\name{mlxUpdateOutput}
\alias{mlxUpdateOutput}
\title{Update output}
\usage{
mlxUpdateOutput(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update output of the structural model by removing previous outputs and adding
penalization.
}
\details{
Should be called after \code{mlxUpdateEquation}
}
\author{
Dominic Bräm
}
