% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlx_update_model.R
\name{mlxUpdatePK}
\alias{mlxUpdatePK}
\title{Update PK section}
\usage{
mlxUpdatePK(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update PK section to define estimated dose levels FX as bioavailability factor
F.
}
\details{
Should be called after \code{mlxUpdateInputs}
}
\author{
Dominic Bräm
}
