% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder.R
\name{newOptiMlx}
\alias{newOptiMlx}
\title{Import structural model}
\usage{
newOptiMlx(pmx_file, lixoft_path)
}
\arguments{
\item{pmx_file}{A Monolix \emph{.mlxtran} file.}

\item{lixoft_path}{If initialized with "Monolix" and the structural model used in the \emph{.mlxtran} file is
a model from the library, \code{lixoft_path} is needed. Standard is "C:/ProgramData/Lixoft/MonolixSuiteXXXX".}
}
\description{
Import the structural model used in a \emph{.mlxtran} run.
}
\author{
Dominic Bräm
}
