% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adder.R
\name{newOptiNlmixr}
\alias{newOptiNlmixr}
\title{Import structural model from nlmixr2}
\usage{
newOptiNlmixr(pmx_file)
}
\arguments{
\item{pmx_file}{A fitted nlmixr2 object.}
}
\description{
Import the structural model used in a nlmixr2-model run.
}
\author{
Dominic Bräm
}
