% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_helpers.R
\name{nlmixr_err_identifier}
\alias{nlmixr_err_identifier}
\title{Identify residual error parameters}
\usage{
nlmixr_err_identifier(fit_obj)
}
\arguments{
\item{fit_obj}{Fitted object with nlmixr2.}
}
\description{
Identify the error parameters in prop(X) and add(X) in the nlmixr2 model.
}
\author{
Dominic Bräm
}
