% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{nlmixrConstraintAUC}
\alias{nlmixrConstraintAUC}
\title{Generate nlmixr2 code for constraint AUC}
\usage{
nlmixrConstraintAUC(
  state,
  limit,
  pen_time,
  eval_time,
  pen_values,
  n_constraint,
  gamma
)
}
\arguments{
\item{state}{State in the structural model to be penalized.}

\item{limit}{The limit, under/above which the AUC gets calculated for penalization.}

\item{pen_time}{Time-range of penalization (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{pen_values}{What values should be penalized; "below" if limit is the lower and "above" if
limit is the upper limit.}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}
}
\description{
Generate nlmixr2 code for constraint, called by \link[=func]{addConstraintAUC} if
software is "nlmixr2".
}
\author{
Dominic Bräm
}
