% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondaries.R
\name{nlmixrSecondaryAUC}
\alias{nlmixrSecondaryAUC}
\title{Generate nlmixr2 code for secondary target AUC}
\usage{
nlmixrSecondaryAUC(
  state,
  pen_values,
  pen_time,
  eval_time,
  n_constraint,
  gamma,
  decay_high = 0.01
)
}
\arguments{
\item{state}{State in the structural model for which AUC is calculated to be penalized.}

\item{pen_values}{What values should be penalized; "high" if high AUCs and "low" if
low AUCs should be penalized.}

\item{pen_time}{Time-range when penalization should be applied (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{eval_time}{Time-point at which the constraint AUC should be evaluated.}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}

\item{decay_high}{Rate of penalization-decay with increasing state-values.}
}
\description{
Generate nlmixr2 code for secondary target, called by \link[=func]{addSecondaryAUC} if
software is "nlmixr2".
}
\details{
Penalization function is just \emph{AUC_state} if high AUCs are penalized, and
exp(-0.01 * AUC_state) if low AUCs are penalized.
}
\author{
Dominic Bräm
}
