% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_setup.R
\name{nlmixrSetAll}
\alias{nlmixrSetAll}
\title{Set up the nlmixr2 optiProject}
\usage{
nlmixrSetAll(
  optiProject,
  name,
  pop,
  iiv,
  sim,
  n_id,
  const_err,
  run_model,
  control
)
}
\arguments{
\item{optiProject}{A optiProject-object with a model, penalizations, dose levels, and parameters.
For requirements regarding penalizations, dose levels, and parameters.}

\item{name}{Name under which optiProject will be saved.}

\item{pop}{Whether multiple subjects should be used (TRUE), either from added individual parameters or
simulated from population parameters, or if only one subject with population parameters should be used
(FALSE).}

\item{iiv}{iiv Whether doses should include inter-individual variability (TRUE) or not (FALSE).}

\item{sim}{If \code{pop=TRUE}, whether individuals should be simulated (TRUE) or added individual
parameters should be used (FALSE).}

\item{n_id}{If \code{pop=TRUE} and \code{sim=TRUE}, number of individuals to be simulated.}

\item{const_err}{If doses are estimated with inter-individual variability for a population of individual
subjects, \code{const_err} defines the balance between fulfilling the objectives (constraints, target values etc.)
and having the doses log-normal distributed.}

\item{run_model}{Whether the generated nlmixr2 optiProject should directly be run.}

\item{control}{If \code{run_model=TRUE}, list of control arguments given to nlmxir2. If \code{iiv=FALSE},
control for "bobyqa" and if \code{iiv=TRUE} control for "saem" must be given.}
}
\description{
Function to generate data and model from an optiProject-object
}
\details{
Data and model will be saved in the global environment, \emph{_data} will be added to the name 
for the data and \emph{_model} to the name for the model.
}
\author{
Dominic Bräm
}
