% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_functions.R
\name{nlmixrTargetFun}
\alias{nlmixrTargetFun}
\title{Generate nlmixr2 code for target function penalization}
\usage{
nlmixrTargetFun(state, target_fun, time, n_constraint, gamma)
}
\arguments{
\item{state}{State in the structural model to be penalized.}

\item{target_fun}{Target function for the state.}

\item{time}{Time-point of penalization (one numeric value).}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}
}
\description{
Generate nlmixr2 code for target function penalization, called by \link[=func]{addTargetFun} if
software is "nlmixr2".
}
\author{
Dominic Bräm
}
