% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_values.R
\name{nlmixrTargetValue}
\alias{nlmixrTargetValue}
\title{Generate nlmixr2 code for target value penalization}
\usage{
nlmixrTargetValue(state, target_value, time, n_constraint, gamma)
}
\arguments{
\item{state}{State in the structural model to be penalized.}

\item{target_value}{Target value of the state.}

\item{time}{Time-point of penalization (one numeric value).}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}
}
\description{
Generate nlmixr2 code for target value penalization, called by \link[=func]{addTargetValue} if
software is "nlmixr2".
}
\author{
Dominic Bräm
}
