% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_values.R
\name{nlmixrTargetValueAUC}
\alias{nlmixrTargetValueAUC}
\title{Generate nlmixr2 code for target value AUC penalization}
\usage{
nlmixrTargetValueAUC(
  state,
  target_value,
  pen_time,
  eval_time,
  n_constraint,
  gamma
)
}
\arguments{
\item{state}{State in the structural model to be penalized.}

\item{target_value}{Target value for the state.}

\item{pen_time}{Time-range when penalization should be applied (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{eval_time}{Time-point at which the constraint AUC should be evaluated.}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}
}
\description{
Generate nlmixr2 code for target value AUC penalization, called by \link[=func]{addTargetValueAUC} if
software is "nlmixr2".
}
\author{
Dominic Bräm
}
