% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_update_model.R
\name{nlmixrUpdateIni}
\alias{nlmixrUpdateIni}
\title{Update ini section of model}
\usage{
nlmixrUpdateIni(optiProject, pop, iiv, const_err)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}

\item{iiv}{Whether doses should contain IIV.}

\item{const_err}{Constant error for error model.}
}
\description{
Remove model parameters and add dosing parameters.
}
\details{
Should be called first.
}
\author{
Dominic Bräm
}
