% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_update_model.R
\name{nlmixrUpdateModelBioav}
\alias{nlmixrUpdateModelBioav}
\title{Update model section bioavailability definition}
\usage{
nlmixrUpdateModelBioav(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Adding bioavailability to the model section.
}
\details{
Should be called after \code{nlmixrUpdateModelParms}
}
\author{
Dominic Bräm
}
