% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_update_model.R
\name{nlmixrUpdateModelInis}
\alias{nlmixrUpdateModelInis}
\title{Update model section initial states}
\usage{
nlmixrUpdateModelInis(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Adding initial states for AUC constraints to the model section.
}
\details{
Should be called after \code{nlmixrUpdateModelInis}
}
\author{
Dominic Bräm
}
