% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_update_model.R
\name{nlmixrUpdateModelOut}
\alias{nlmixrUpdateModelOut}
\title{Update model section output code}
\usage{
nlmixrUpdateModelOut(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Removing previous model output and adding penalization output to the model section.
}
\details{
Should be called after \code{nlmixrUpdateModelPens}
}
\author{
Dominic Bräm
}
