% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_update_model.R
\name{nlmixrUpdateModelParms}
\alias{nlmixrUpdateModelParms}
\title{Update model section dose definition}
\usage{
nlmixrUpdateModelParms(optiProject, iiv)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}

\item{iiv}{Whether doses should contain IIV.}
}
\description{
Adding dose definitions to the model section.
}
\details{
Should be called after \code{nlmixrUpdateModelInis}
}
\author{
Dominic Bräm
}
