% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_update_model.R
\name{nlmixrUpdateModelPens}
\alias{nlmixrUpdateModelPens}
\title{Update model section penalisation code}
\usage{
nlmixrUpdateModelPens(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Adding penalisation code to the model section.
}
\details{
Should be called after \code{nlmixrUpdateModelBioav}
}
\author{
Dominic Bräm
}
