% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{nm_multi_sep_code}
\alias{nm_multi_sep_code}
\title{Generate code for multiple seperate subjects in NONMEM}
\usage{
nm_multi_sep_code(optiProject, name, save_path)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{name}{Name under which the optiProject is saved.}

\item{save_path}{Path where optiProject is saved.}
}
\description{
To estimate multiple subject seperately in NONMEM $PROB, $INPUT, $DATA, 
$THETA, $EST, and $TABLE can be saved in a seperate file and included in the
control file
}
\author{
Dominic Bräm
}
