% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{nm_nstate_extractor}
\alias{nm_nstate_extractor}
\title{Get maximal number of compartments}
\usage{
nm_nstate_extractor(model_text)
}
\arguments{
\item{model_text}{Structural NONMEM model.}
}
\description{
Get the maximal number of compartments in the structural model. Needed to define
the compartment number for penalizations with AUC.
}
\details{
Looking for highest DADT(X) in the model.
}
\author{
Dominic Bräm
}
