% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{nmConstraintAUC}
\alias{nmConstraintAUC}
\title{Generate NONMEM code for constraint AUC}
\usage{
nmConstraintAUC(
  state,
  limit,
  pen_time,
  eval_time,
  pen_values,
  n_state,
  n_constraint,
  gamma
)
}
\arguments{
\item{state}{State in the structural model to be penalized.}

\item{limit}{The limit, under/above which the AUC gets calculated for penalization.}

\item{pen_time}{Time-range of penalization (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{pen_values}{What values should be penalized; "below" if limit is the lower and "above" if
limit is the upper limit.}

\item{n_state}{The index of the state for AUC (number of states in the structural model plus 1)}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}
}
\description{
Generate NONMEM code for constraint, called by \link[=func]{addConstraintAUC} if
software is "NONMEM".
}
\author{
Dominic Bräm
}
