% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_setup.R
\name{nmSetAll}
\alias{nmSetAll}
\title{Set up the NONMEM optiProject}
\usage{
nmSetAll(optiProject, name, save_path, pop, iiv, sim, n_id, separate)
}
\arguments{
\item{optiProject}{A optiProject-object with a model, penalizations, dose levels, and parameters.
For requirements regarding penalizations, dose levels, and parameters.}

\item{name}{Name under which optiProject will be saved. .ctl will be added to name.}

\item{save_path}{Path, where optiProject should be saved.}

\item{pop}{Whether multiple subjects should be used (TRUE), either from added individual parameters or
simulated from population parameters, or if only one subject with population parameters should be used
(FALSE).}

\item{iiv}{iiv Whether doses should include inter-individual variability (TRUE) or not (FALSE).}

\item{sim}{If \code{pop=TRUE}, whether individuals should be simulated (TRUE) or added individual
parameters should be used (FALSE).}

\item{n_id}{If \code{pop=TRUE} and \code{sim=TRUE}, number of individuals to be simulated.}

\item{seperate}{If \code{pop=TRUE} and \code{iiv=FALSE}, whether individuals should be fitted separately (TRUE) or
all individuals should be fitted together.}
}
\description{
Function to generate data and model from an optiProject-object
}
\author{
Dominic Bräm
}
