% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_functions.R
\name{nmTargetFunAUC}
\alias{nmTargetFunAUC}
\title{Generate NONMEM code for target function AUC penalization}
\usage{
nmTargetFunAUC(
  state,
  target_fun,
  pen_time,
  eval_time,
  n_state,
  n_constraint,
  gamma
)
}
\arguments{
\item{state}{State in the structural model to be penalized.}

\item{target_fun}{Target function for the state.}

\item{pen_time}{Time-range when penalization should be applied (numeric vector of length 2, i.e., c(min_time, max_time)).}

\item{eval_time}{Time-point at which the constraint AUC should be evaluated.}

\item{n_state}{The index of the state for AUC (number of states in the structural model plus 1).}

\item{n_constraint}{The index of the penalization (number of penalizations in the optiProject-object 
plus 1)}

\item{gamma}{Penalization strength.}
}
\description{
Generate NONMEM code for target function AUC penalization, called by \link[=func]{addTargetFunAUC} if
software is "NONMEM".
}
\author{
Dominic Bräm
}
