% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateDES}
\alias{nmUpdateDES}
\title{Update $DES in a NONMEM model}
\usage{
nmUpdateDES(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update the $DES section in a NONMEM model, adding the constraint/target-codes
to the $DES section
}
\details{
Should be called after \code{nmUpdatePK}.
}
\author{
Dominic Bräm
}
