% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateError}
\alias{nmUpdateError}
\title{Update $ERROR in a NONMEM model}
\usage{
nmUpdateError(optiProject, iiv = FALSE)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}

\item{iiv}{Whether inter-individual variability should be added to the estimated doses.}
}
\description{
Update the $ERROR section in a NONMEM model, removing previous $ERROR code and 
adding $ERROR code for penalization.
}
\details{
Should be called after \code{nmUpdateDES}.
}
\author{
Dominic Bräm
}
