% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateEst}
\alias{nmUpdateEst}
\title{Update $EST in a NONMEM model}
\usage{
nmUpdateEst(optiProject, iiv = FALSE)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}

\item{iiv}{Whether inter-individual variability should be added to the estimated doses.}
}
\description{
Update the $EST line in a NONMEM model, removing previous $EST lines and adding
$EST line for optiProject (with iiv not yet implemented).
}
\details{
If model is with IIV, SAEM will be used. Should be called after \code{nmUpdateSigma}.
}
\author{
Dominic Bräm
}
