% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateInputs}
\alias{nmUpdateInputs}
\title{Update $INPUT in a NONMEM model}
\usage{
nmUpdateInputs(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update the $INPUT line in a NONMEM model
}
\details{
Should be called after \code{nmUpdateThetaEta}.
}
\author{
Dominic Bräm
}
