% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateMulti}
\alias{nmUpdateMulti}
\title{Update NONMEM model for multiple seperate individuals}
\usage{
nmUpdateMulti(optiProject, name, sim, n_id)
}
\arguments{
\item{optiProject}{A optiProject-object}

\item{name}{Name under which the optiProject is saved.}

\item{sim}{Whether individuals were simulated (TRUE) or IParms was used (FALSE).}

\item{n_id}{Number of simulated individuals.}
}
\description{
RECS=ID is added to the model $DATA line and INCLUDE of the multiple fitting
text file at the end of the model.
}
\author{
Dominic Bräm
}
