% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateNcomp}
\alias{nmUpdateNcomp}
\title{Update NCOMPARTMENTS in a NONMEM model}
\usage{
nmUpdateNcomp(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update the NCOMPARTMENTS argument in a NONMEM model.
}
\details{
Should be called after \code{nmUpdateSubrout}.
}
\author{
Dominic Bräm
}
