% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateOmega}
\alias{nmUpdateOmega}
\title{Update $OMEGA in a NONMEM model}
\usage{
nmUpdateOmega(optiProject, iiv = FALSE)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}

\item{iiv}{Whether inter-individual variability should be added to the estimated doses.}
}
\description{
Update the $OMEGA section in a NONMEM model, either removing $OMEGA section
entirely if \code{iiv=FALSE}, or removing previous $OMEGA code and adding
initial values for dose levels.
}
\details{
Should be called after \code{nmUpdateTheta}.
}
\author{
Dominic Bräm
}
