% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdatePK}
\alias{nmUpdatePK}
\title{Update $PK in a NONMEM model}
\usage{
nmUpdatePK(optiProject, iiv = FALSE)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}

\item{iiv}{Whether inter-individual variability should be added to the estimated doses.}
}
\description{
Update the $PK line in a NONMEM model, defining the dose levels as THETAs and ETAs.
}
\details{
If model is generated with IIV, MU-referencing will be applied.
Should be called after \code{nmUpdateNcomp}.
}
\author{
Dominic Bräm
}
