% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateSigma}
\alias{nmUpdateSigma}
\title{Update $SIGMA in a NONMEM model}
\usage{
nmUpdateSigma(optiProject, iiv = FALSE)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}

\item{iiv}{Whether inter-individual variability should be added to the estimated doses.}
}
\description{
Update the $SIGMA section in a NONMEM model, either removing $SIGMA section
entirely if \code{iiv=FALSE}, or removing previous $SIGMA code and adding
initial values for residual error parameter (not yet implemented).
}
\details{
Should be called after \code{nmUpdateOmega}.
}
\author{
Dominic Bräm
}
