% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateSubrout}
\alias{nmUpdateSubrout}
\title{Update $SUBROUTINES in a NONMEM model}
\usage{
nmUpdateSubrout(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update the $SUBROUTINGES line in a NONMEM model.
}
\details{
Should be called after \code{nmUpdateData}. Tolerances will be set to 15.
}
\author{
Dominic Bräm
}
