% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateTable}
\alias{nmUpdateTable}
\title{Update $TABLE in a NONMEM model}
\usage{
nmUpdateTable(optiProject, name, iiv = FALSE)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}

\item{iiv}{Whether inter-individual variability should be added to the estimated doses.}
}
\description{
Update the $TABLE line in a NONMEM model, removing previous $TABLE lines and adding
outputs TIME PEN
}
\details{
Should be called after \code{nmUpdateEst}.
}
\author{
Dominic Bräm
}
