% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateTheta}
\alias{nmUpdateTheta}
\title{Update $THETA in a NONMEM model}
\usage{
nmUpdateTheta(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update the $THETA section in a NONMEM model, removing previous $THETA code and
adding initial values for dose levels.
}
\details{
Should be called after \code{nmUpdateError}.
}
\author{
Dominic Bräm
}
