% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_update_model.R
\name{nmUpdateThetaEta}
\alias{nmUpdateThetaEta}
\title{Update Thetas and Etas in Model}
\usage{
nmUpdateThetaEta(optiProject)
}
\arguments{
\item{optiProject}{A optiProject-object with a model added.}
}
\description{
Update THETA(X) and ETA(Y) to THETAX_cov and ETAY_cov in the NONMEM model.
}
\details{
Should be called first to avoid updating dose THETAs and ETAs.
}
\author{
Dominic Bräm
}
