% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{omega_extractor_nm}
\alias{omega_extractor_nm}
\title{Omega estimates extractor for NONMEM}
\usage{
omega_extractor_nm(res_path)
}
\arguments{
\item{res_path}{NONMEM results file.}
}
\description{
Extracts Omega estimates from a NONMEM results file.
}
\author{
Dominic Bräm
}
