% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{parm_extractor}
\alias{parm_extractor}
\title{General estimates extractor}
\usage{
parm_extractor(pmx_file, software, phi_file = NULL)
}
\arguments{
\item{pmx_file}{Name of either a Monolix .mlxtran file or a NONMEM results file.}

\item{software}{Initialized software of the optiProject-object.}

\item{phi_file}{If \code{software="NONMEM"}, a .phi file of the NONMEM run to extract
individual etas.}
}
\description{
Gets estimated population and individual parameters from a Monolix or NONMEM run.
}
\author{
Dominic Bräm
}
