% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{parm_extractor_mlx}
\alias{parm_extractor_mlx}
\title{Population estimates extractor for Monolix}
\usage{
parm_extractor_mlx(parm_file)
}
\arguments{
\item{parm_file}{File of estimated population parameters.}
}
\description{
Extracts population estimates from a Monolix run.
}
\author{
Dominic Bräm
}
