% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{parm_extractor_nlmixr}
\alias{parm_extractor_nlmixr}
\title{Population estimates extractor for nlmixr2}
\usage{
parm_extractor_nlmixr(fit_obj, err_parms)
}
\arguments{
\item{fit_obj}{Fitted object with nlmixr2.}
}
\description{
Extracts population estimates from a nlmixr2 run.
}
\author{
Dominic Bräm
}
