% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{parm_extractor_nm}
\alias{parm_extractor_nm}
\title{Population estimates extractor for NONMEM}
\usage{
parm_extractor_nm(res_file)
}
\arguments{
\item{res_file}{NONMEM results file.}
}
\description{
Extracts population estimates from a NONMEM results file.
}
\author{
Dominic Bräm
}
