% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{PParmDataGen}
\alias{PParmDataGen}
\title{Generate parameter data for typical subject}
\usage{
PParmDataGen(parms, software)
}
\arguments{
\item{parms}{Population parameter data frame with columns for parameter name and value,
including at least \emph{_pop} parameters for Monolix or \emph{THETA} parameters for NONMEM}

\item{software}{Software of the optiProject-object}
}
\description{
Generate  parameter data for the typcial subject from the population information of
an optiProject-object
}
\author{
Dominic Bräm
}
