% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{theta_extractor_nlmixr}
\alias{theta_extractor_nlmixr}
\title{Fixed effects extractor for nlmixr2}
\usage{
theta_extractor_nlmixr(fit_obj, err_parms)
}
\arguments{
\item{fit_obj}{Fitted object with nlmixr2.}

\item{err_parms}{Names of residual error parameters in the model, extracted
through \link[=func]{nlmixr_err_identifier} to be excluded from fixed effects.}
}
\description{
Extract estimated fixed effects from nlmixr2 fit.
}
\author{
Dominic Bräm
}
