% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parm_extractors.R
\name{theta_extractor_nm}
\alias{theta_extractor_nm}
\title{Theta estimates extractor for NONMEM}
\usage{
theta_extractor_nm(res_path)
}
\arguments{
\item{res_path}{NONMEM results file.}
}
\description{
Extracts Theta estimates from a NONMEM results file.
}
\author{
Dominic Bräm
}
