% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_helpers.R
\name{update_nm_model_parms}
\alias{update_nm_model_parms}
\title{Update Thetas and Etas in Model}
\usage{
update_nm_model_parms(model_text)
}
\arguments{
\item{model_text}{A structural NONMEM model.}
}
\description{
Update THETA(X) and ETA(Y) to THETAX_cov and ETAY_cov in the NONMEM model.
}
\details{
Should be called first to avoid updating dose THETAs and ETAs.
}
\author{
Dominic Bräm
}
